chore: Format entities API to IASO structure
